Got project building on Linux with new CMake files; Fixed a few warnings
git-svn-id: https://hypertable.googlecode.com/svn/trunk@28 01261837-d133-0410-8dcb-6792eb736156

+++ src/cc/Hypertable/RangeServer/CommitLogReaderLocal.cc
@@ -100,7 +100,7 @@ CommitLogReaderLocal::CommitLogReaderLocal(std::string &logDirRoot, std::string 
       exit(1);
     }
 
-    if (statbuf.st_size < sizeof(CommitLogHeaderT))
+    if (statbuf.st_size < (off_t)sizeof(CommitLogHeaderT))
       continue;
 
     if ((fd = open(mLogFileInfo[i].fname.c_str(), O_RDONLY)) == -1) {
